+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/keys/DeleteRoomKeysResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DeleteRoomKeysResponse(
     @SerialName("count")
-    val count: Int,
+    val count: Long,
     @SerialName("etag")
     val etag: String
 )
